requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.18.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
requests>=2.31.0
rapidfuzz>=3.6.0
beautifulsoup4>=4.12.0
